Add `VideoNotOwnedWarning` component
